// src/interfaces/song.ts
export interface Song {
    id:            string;
    dz_Id:         number;
    title:         string;
    artist:        string;
    preview:       string;
    image_small:   string;
    image_medium:  string;
    image_big:     string;
    reproductions: number;
    duration:      number;
    genre:         string;
    album:         string;
    type:          string;
}